Added Order functonality:
Added functionality to place an Order. Requires bikeId userId and amount of money. Order is created and status of bike is updated. Need to add ability to deduct balance from student card when this functionality is in place.

+++ springBootDemo/src/main/java/ie/demo/service/impl/CustomerFactory.java
@@ -9,7 +9,7 @@ import ie.demo.service.UserFactory;
 public class CustomerFactory implements UserFactory {
 
 	@Override
-	public User createUser(String username, String password, long studentCardId, String email) {
+	public User createUser(String username, String password, int studentCardId, String email) {
 		return new User(username, password, studentCardId, email, 1);
 	}
